Se agregaron clases que permiten ordenar los resultados de las busquedas para los expertos, además se modificaron las clases necesarias para realizar la integración con la pasarela de pagos Epayco

+++ src/main/java/com/espiritware/opusclick/controller/ProviderController.java
@@ -11,7 +11,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -30,6 +29,8 @@ import com.espiritware.opusclick.model.State;
 import com.espiritware.opusclick.model.Work;
 import com.espiritware.opusclick.service.AmazonClient;
 import com.espiritware.opusclick.service.ProviderService;
+import com.espiritware.opusclick.util.ReverseSortProviderByReputation;
+import com.espiritware.opusclick.util.ReverseSortProviderReviewByDate;
 
 @Controller
 //@CrossOrigin(origins = {"http://localhost:4200"}, maxAge = 4800, allowCredentials = "false")
@@ -37,6 +38,9 @@ import com.espiritware.opusclick.service.ProviderService;
 public class ProviderController {
 	
 	private static final String PROVIDER_IMAGES_FOLDER="provider-profile-images/";
+	
+	private static final String PROVIDER_DEFAULT_IMAGE="https://s3.amazonaws.com/opusclick.com/provider-profile-images/default-profile-photo.png";
+
 	
 	@Autowired
 	private ProviderService providerService;
@@ -72,7 +76,7 @@ public class ProviderController {
 			if (providers.isEmpty()) {
 				return new ResponseEntity<List<?>>(providers, HttpStatus.OK);
 			} else {
-				Collections.sort(providers);
+				providers.sort(new ReverseSortProviderByReputation());
 				List<ProviderGetByProfessionDto> dtoList=getProviderDtoList(providers);
 				return new ResponseEntity<List<?>>(dtoList, HttpStatus.OK);
 			}
@@ -81,6 +85,7 @@ public class ProviderController {
 			if (providers.isEmpty()) {
 				return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 			} else {
+				providers.sort(new ReverseSortProviderByReputation());
 				return new ResponseEntity<List<?>>(providers, HttpStatus.OK);
 			}
 		}
@@ -159,6 +164,7 @@ public class ProviderController {
 				}
 			}
 		}
+		dtoList.sort(new ReverseSortProviderReviewByDate());
 		return dtoList;
 	}
 	
@@ -197,7 +203,10 @@ public class ProviderController {
 		if (provider.getPhoto() != null) {
 			if (!provider.getPhoto().isEmpty()) {
 				try {
-					amazonClient.deleteFileFromS3Bucket(PROVIDER_IMAGES_FOLDER,provider.getPhoto());
+					if(!provider.getPhoto().equalsIgnoreCase(PROVIDER_DEFAULT_IMAGE)) {
+						amazonClient.deleteFileFromS3Bucket(PROVIDER_IMAGES_FOLDER,provider.getPhoto());
+					}
+					
 				} catch (Exception e) {
 					return new ResponseEntity<>(
 							new CustomErrorType("Provider with id: " + providerId + " can't be erased"),
